+++ config/pipeline_config.json
@@ -1,4 +1,5 @@
 {
+  "data_type": "synthetic",
   "input_parameters": "parameters_pulmonary_sensitive_summarystats.json",
   "output_path": "output_synthetic/",
   "steps": ["1","2","3","4","5", "6", "7"],  
@@ -10,7 +11,8 @@
     "t_max_dpdt", "a_epad", "epad", "s_a_epad", "s_epad",
     "min_dpdt", "max_dpdt", "A_p", "P_max", "esp", "sys",
     "EF", "Ees/Ea", "iT", "PC1", "PC2", "PC3", "MC_EF"
+  ],
   "include_timeseries": 0,
   "epsilon_obs_scale": 0.05,
-  "output_dir_bayesian": "output_synthetic/output_5000_9params/bayesian_calibration_results/18_output_keys/calibration_20250619_135107"
+  "output_dir_bayesian": "output_synthetic/output_5000_9_params/bayesian_calibration_results/18_output_keys/calibration_20250619_135107"
 }
